+++ src/app/common/components/components.module.ts
@@ -2,18 +2,41 @@ import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
 import {SpinnerComponent} from "./spinner/spinner.component";
 import {MatProgressSpinnerModule} from "@angular/material/progress-spinner";
-
+import {ToolbarComponent} from './toolbar/toolbar.component';
+import {MatToolbarModule} from "@angular/material/toolbar";
+import {MatIconModule} from "@angular/material/icon";
+import {MatButtonModule} from "@angular/material/button";
+import {MatMenuModule} from "@angular/material/menu";
+import {RouterLink, RouterLinkActive} from "@angular/router";
+import {MatTabsModule} from "@angular/material/tabs";
+import {MatCardModule} from "@angular/material/card";
+import {CardComponent} from './card/card.component';
+import {HeaderComponent} from "./header/header.component";
 
 @NgModule({
   declarations: [
-    SpinnerComponent
+    SpinnerComponent,
+    ToolbarComponent,
+    CardComponent,
+    HeaderComponent
   ],
   exports: [
-    SpinnerComponent
+    SpinnerComponent,
+    ToolbarComponent,
+    CardComponent,
+    HeaderComponent
   ],
   imports: [
     CommonModule,
-    MatProgressSpinnerModule
+    MatProgressSpinnerModule,
+    MatToolbarModule,
+    MatIconModule,
+    MatButtonModule,
+    MatMenuModule,
+    RouterLink,
+    MatTabsModule,
+    MatCardModule,
+    RouterLinkActive,
   ]
 })
 export class ComponentsModule {