+++ src/app/features/owner/update-owner/update-owner.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-update-owner',
+  templateUrl: './update-owner.component.html',
+  styleUrls: ['./update-owner.component.scss']
+})
+export class UpdateOwnerComponent {
+
+}