+++ src/app/features/property/properties-dashboard/properties-dashboard.component.ts
@@ -1,0 +1,11 @@
+import {Component} from '@angular/core';
+
+@Component({
+  selector: 'app-property',
+  templateUrl: './properties-dashboard.component.html',
+  styleUrls: ['./properties-dashboard.component.scss']
+})
+export class PropertiesDashboardComponent {
+  propertyId: string = "12345"; // TODO get ID
+
+}