+++ src/app/features/tenant/create-tenant/create-tenant.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-create-tenant',
+  templateUrl: './create-tenant.component.html',
+  styleUrls: ['./create-tenant.component.scss']
+})
+export class CreateTenantComponent {
+
+}